chore: move test template
